+++ testsuite/tests/dependent/should_fail/T16326_Fail10.hs
@@ -1,0 +1,12 @@
+{-# LANGUAGE PolyKinds #-}
+{-# LANGUAGE RankNTypes #-}
+module T16326_Fail10 where
+
+import Data.Kind
+
+type Const a b = a
+
+flurmp :: a -> ()
+flurmp _ = ()
+{-# RULES "flurmp"
+    forall (x :: forall a -> a -> a). flurmp x = () #-}