+++ testsuite/tests/dependent/should_fail/T16326_Fail2.hs
@@ -1,0 +1,6 @@
+{-# LANGUAGE ForeignFunctionInterface #-}
+{-# LANGUAGE PolyKinds #-}
+{-# LANGUAGE RankNTypes #-}
+module T16326_Fail2 where
+
+foreign import ccall "blah" blah :: forall a -> a -> IO ()